Octobre 1 2021

+++ app/src/main/java/com/begemot/myapplicationz/Cache.kt
@@ -1,7 +1,7 @@
 package com.begemot.myapplicationz
 
 import android.graphics.BitmapFactory
-import android.util.Log
+import android.text.format.Formatter
 import androidx.compose.runtime.MutableState
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.asImageBitmap
@@ -13,6 +13,8 @@ import java.text.SimpleDateFormat
 import java.util.stream.Collectors
 //import kotlin.io.path.ExperimentalPathApi
 import kotlin.io.path.fileSize
+import kotlin.io.path.getLastModifiedTime
+
 //import kotlin.io
 
 private val DD=false
@@ -26,21 +28,45 @@ object KCache{
 
         try {
             Files.write(Paths.get(sf+"/$sNameFile"),scontent.toByteArray())
-            Files.write(Paths.get(sf+"/ERROR.BGM"),"ok namefile:$sNameFile  size ${scontent.length}\n".toByteArray(),StandardOpenOption.APPEND,StandardOpenOption.CREATE,StandardOpenOption.WRITE)
+            //Files.write(Paths.get(sf+"/ERROR.BGM"),"ok namefile:$sNameFile  size ${scontent.length}\n".toByteArray(),StandardOpenOption.APPEND,StandardOpenOption.CREATE,StandardOpenOption.WRITE)
         } catch (e: Exception) {
-            Files.write(Paths.get(sf+"/ERROR.BGM"),"err ${e.message}\n".toByteArray(),StandardOpenOption.APPEND,StandardOpenOption.CREATE,StandardOpenOption.WRITE)
-
+            //Files.write(Paths.get(sf+"/ERROR.BGM"),"err ${e.message}\n".toByteArray(),StandardOpenOption.APPEND,StandardOpenOption.CREATE,StandardOpenOption.WRITE)
+            writeError("storeInCache $sNameFile ${e.message}")
         }
     }
 
-    fun storeImageInCache(sNameFile: String, zimage: ByteArray){
+    fun writeError(serr:String){
+        val  sf= App.lcontext.filesDir.absolutePath
+
+        Files.write(Paths.get(sf+"/ERROR.BGM"),"err ${serr}\n".toByteArray(),StandardOpenOption.APPEND,StandardOpenOption.CREATE,StandardOpenOption.WRITE)
+
+    }
+
+    fun readErrors():String{
+        val  sf= App.lcontext.filesDir.absolutePath
+        if(!fileExistsAndNotEmpty("/ERROR.BGM","")) return "no errors logged"
+        return String(Files.readAllBytes(Paths.get(sf+"/ERROR.BGM")))
+        //return Files.readAllLines(Paths.get(sf+"/ERROR.BGM")).joinToString { "\n" }
+    }
+
+    fun clearErrors(){
+        val  sf= App.lcontext.filesDir.absolutePath
+        try {
+            Files.delete(Paths.get("$sf/ERROR.BGM"))
+        } catch (e: Exception) {
+        }
+    }
+
+
+    fun storeImageInFile(sNameFile: String, zimage: ByteArray){
         Timber.d("STORE IMAGE IN CACHE $sNameFile   with size ${zimage.size}")
+        if(zimage.size==0) return
         val  sf=App.lcontext.filesDir.absolutePath + "/Images/$sNameFile"
         Files.write(Paths.get(sf),zimage)
         Timber.d("end write $sf size ${zimage.size}")
     }
 
-    suspend fun loadFromCache(sNameFile: String):String= withContext(Dispatchers.IO){
+     fun loadFromCache(sNameFile: String):String{   //= withContext(Dispatchers.IO){
         val  sf="${App.lcontext.filesDir.absolutePath}/$sNameFile"
         val str=
         try {
@@ -50,7 +76,7 @@ object KCache{
             ""//"error loadFromCache ${e.message}"
         }
         Timber.d("File raw $sf ->' ${str.substring(0,minOf(str.length,29))}....'")
-        str
+        return str
     }
 
    fun setUp(){
@@ -75,17 +101,29 @@ object KCache{
         return file.exists()
     }
 
-    fun findImageInFile(sNameFile: String):ByteArray{
-        val zimage:ByteArray
+    fun addImageInMemory(sNameImg:String,img:ImageBitmap){
+        hashmapImages[sNameImg] = img
+    }
+
+    fun findImageInFile(sNameFile: String):ImageBitmap? {
+        /* val l=findImageInFile( sNameImg)
+        Timber.d("$sNameImg    size ${l.size}")
+        if(l.size>0) {
+            val s = BitmapFactory.decodeByteArray(l, 0, l.size).asImageBitmap()
+            hashmapImages[sNameImg] = s
+             s
+        }*/
+        //val zimage:ByteArray
         try {
             val  sf=App.lcontext.filesDir.absolutePath + "/Images/${sNameFile}"
-            zimage = Files.readAllBytes(Paths.get(sf))
-
+            val zimage = Files.readAllBytes(Paths.get(sf))
+            val s = BitmapFactory.decodeByteArray(zimage, 0, zimage.size).asImageBitmap()
+            return s
         } catch (e: Exception) {
             Timber.d("cache exception finding $sNameFile $e")
-            return  ByteArray(0)
+            return  null //ByteArray(0)
         }
-        return  zimage
+        //return  zimage
     }
 
     fun getBitmapImage2(sNameImg: String, ms: MutableState<ImageBitmap>){
@@ -93,29 +131,34 @@ object KCache{
         scope.launch {
             //Log.d("UPALA","hola")
                 Timber.d("$sNameImg")
-                val x=getBitmapImage(sNameImg)
+                val x=getBitmapImageFromMemCache(sNameImg)
                 if(x != null) ms.value = x
                 else Timber.d("Image Null $sNameImg")
         }
      }
 
 
-   suspend  fun getBitmapImage(sNameImg: String): ImageBitmap?= withContext(Dispatchers.IO) {
-        val bmp=hashmapImages[sNameImg]
+   suspend  fun getBitmapImageFromMemCache(sNameImg: String): ImageBitmap?= withContext(Dispatchers.IO) {
+        return@withContext hashmapImages[sNameImg]
        //Timber.d("get BitmapImage")
-        if(bmp!=null) {
-            return@withContext bmp
-        }
-        val l=findImageInFile( sNameImg)
+        //if(bmp!=null) {
+        //    return@withContext bmp
+        //}
+       /* val l=findImageInFile( sNameImg)
         Timber.d("$sNameImg    size ${l.size}")
         if(l.size>0) {
             val s = BitmapFactory.decodeByteArray(l, 0, l.size).asImageBitmap()
             hashmapImages[sNameImg] = s
              s
-        }
-        else  null
+        }*/
+        //else  null
     }
 
+    /*fun filesInDir(dir:String){
+        val  sf="${App.lcontext.filesDir.absolutePath}/$dir"
+        val f=Files.list(Paths.get(sf))
+        f.filter(  )
+    }*/
 
 
 
@@ -126,6 +169,7 @@ object KCache{
             if(f==0L) return false
             return File(sf).exists()
         } catch (e: Exception) {
+            Timber.d("$e")
             return false  //Mes pos i no cagues
         }
     }
@@ -152,6 +196,25 @@ fun listAllFiles():List<String>{
 
         val ls= result.map{it->"${it.toString().substring(root.length)}  ${it?.fileSize()}"}
         return ls
+    }
+
+    fun listDirectory(sDir:String):String{
+        val path=Paths.get("${App.lcontext.filesDir.absolutePath}/$sDir")
+        val fd = Files.list(path).collect(Collectors.toList())
+        //val s=StringBuilder()
+        val l= mutableListOf("")
+        //DateFormat
+        val df = SimpleDateFormat("dd/MM/yyyy")
+
+        fd.forEach {
+            val size=Formatter.formatShortFileSize(App.lcontext,it.fileSize())
+            l.add("${it.fileName} (${size}) ${df.format(it.getLastModifiedTime().toMillis())}\n")
+            //s.append("(${it.fileSize().toString().padStart(6,'0')}) ${it.fileName}\n")
+        }
+            //.map(Path::getFileName)
+            //.map(Path::toString)
+        l.sort()
+        return l.joinToString("") //s.toString()
     }
 
 
@@ -192,8 +255,11 @@ fun listAllFiles():List<String>{
         val F=File(App.lcontext.filesDir.absolutePath+"/$dir")
         F.listFiles()?.filter { it.isFile }
             ?.forEach{
-                Timber.d("Remove : ${it.name}")
-                it.delete()
+                Timber.d("Remove : ${it.name}  ${it.extension}")
+                if(!it.extension.equals("BKM") && !it.extension.equals("LIN")) {        //do not remove bookmarks
+                    Timber.d("RELLY Remove : ${it.name}")
+                    it.delete()
+                }
             }
     }
